ADT-449 Rename getStatusCommit file to getLastCommit
